+++ projects/ngx-skeleton-loader/src/lib/ngx-skeleton-loader.component.spec.ts
@@ -39,6 +39,10 @@ import { NgxSkeletonLoaderComponent } from './ngx-skeleton-loader.component';
         <ngx-skeleton-loader animation="invalid-option"></ngx-skeleton-loader>
       </div>
 
+      <div class="skeletons-count-invalid-option">
+        <ngx-skeleton-loader [count]="invalidValueIncount"></ngx-skeleton-loader>
+      </div>
+
       <div class="skeletons-with-count">
         <ngx-skeleton-loader count="2"></ngx-skeleton-loader>
       </div>
@@ -56,6 +60,7 @@ import { NgxSkeletonLoaderComponent } from './ngx-skeleton-loader.component';
 })
 class ContainerComponent {
   animationWithFalsePassedViaBinding = false;
+  invalidValueIncount = 'two';
 }
 
 describe('NgxSkeletonLoaderComponent', () => {
@@ -73,21 +78,31 @@ describe('NgxSkeletonLoaderComponent', () => {
     }),
   );
 
+  it('should console 2 errors if `animation` and `count` receives invalid options and is running in development mode', () => {
+    expect(console.error).toHaveBeenCalledTimes(2);
+  });
+
   it('should console errors if `animation` is an invalid option and is running in development mode', () => {
-    expect(console.error).toHaveBeenCalledTimes(1);
     expect(console.error).toHaveBeenCalledWith(
       // tslint:disable-next-line: max-line-length
       `\`NgxSkeletonLoaderComponent\` need to receive 'animation' as: progress, progress-dark, pulse, false. Forcing default to "progress".`,
     );
   });
 
+  it('should console errors if `animation` is an invalid option and is running in development mode', () => {
+    expect(console.error).toHaveBeenCalledWith(
+      // tslint:disable-next-line: max-line-length
+      `\`NgxSkeletonLoaderComponent\` need to receive 'count' a numeric value. Forcing default to "1".`,
+    );
+  });
+
   it('should add all relevant WAI-ARIA `aria-` attributes in all ngx-skeleton-loader', () => {
-    expect(fixture.nativeElement.querySelectorAll('[aria-busy="true"]').length).toBe(12);
-    expect(fixture.nativeElement.querySelectorAll('[aria-valuemin="0"]').length).toBe(12);
-    expect(fixture.nativeElement.querySelectorAll('[aria-valuemax="100"]').length).toBe(12);
-    expect(fixture.nativeElement.querySelectorAll('[aria-valuetext]').length).toBe(12);
-    expect(fixture.nativeElement.querySelectorAll('[role="progressbar"]').length).toBe(12);
-    expect(fixture.nativeElement.querySelectorAll('[tabindex="0"]').length).toBe(12);
+    expect(fixture.nativeElement.querySelectorAll('[aria-busy="true"]').length).toBe(13);
+    expect(fixture.nativeElement.querySelectorAll('[aria-valuemin="0"]').length).toBe(13);
+    expect(fixture.nativeElement.querySelectorAll('[aria-valuemax="100"]').length).toBe(13);
+    expect(fixture.nativeElement.querySelectorAll('[aria-valuetext]').length).toBe(13);
+    expect(fixture.nativeElement.querySelectorAll('[role="progressbar"]').length).toBe(13);
+    expect(fixture.nativeElement.querySelectorAll('[tabindex="0"]').length).toBe(13);
   });
 
   it('should use progress as default animation if `animation` is not passed as component attribute', () => {