+++ projects/ngx-skeleton-loader/src/lib/ngx-skeleton-loader.component.ts
@@ -1,4 +1,14 @@
-import { Component, OnInit, Input, isDevMode, OnDestroy, AfterViewInit, ChangeDetectionStrategy } from '@angular/core';
+import {
+  Component,
+  OnInit,
+  Input,
+  isDevMode,
+  OnDestroy,
+  AfterViewInit,
+  ChangeDetectionStrategy,
+  OnChanges,
+  SimpleChanges,
+} from '@angular/core';
 import { start, end } from 'perf-marks/marks';
 
 @Component({
@@ -7,7 +17,7 @@ import { start, end } from 'perf-marks/marks';
   styleUrls: ['./ngx-skeleton-loader.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class NgxSkeletonLoaderComponent implements OnInit, AfterViewInit, OnDestroy {
+export class NgxSkeletonLoaderComponent implements OnInit, AfterViewInit, OnDestroy, OnChanges {
   // tslint:disable-next-line: variable-name
   static ngAcceptInputType_count: number | string;
   // tslint:disable-next-line: variable-name
@@ -37,7 +47,23 @@ export class NgxSkeletonLoaderComponent implements OnInit, AfterViewInit, OnDest
     start('NgxSkeletonLoader:Rendered');
     start('NgxSkeletonLoader:Loaded');
 
-    this.items.length = +this.count;
+    this.validateInputValues();
+  }
+
+  private validateInputValues() {
+    // Checking if it's receiving a numeric value (string having ONLY numbers or if it's a number)
+    if (!/^\d+$/.test(`${this.count}`)) {
+      // Shows error message only in Development
+      if (isDevMode()) {
+        console.error(
+          `\`NgxSkeletonLoaderComponent\` need to receive 'count' a numeric value. Forcing default to "1".`,
+        );
+      }
+      this.count = 1;
+    }
+
+    this.items.length = this.count;
+
     const allowedAnimations = ['progress', 'progress-dark', 'pulse', 'false'];
     if (allowedAnimations.indexOf(String(this.animation)) === -1) {
       // Shows error message only in Development
@@ -50,6 +76,22 @@ export class NgxSkeletonLoaderComponent implements OnInit, AfterViewInit, OnDest
       }
       this.animation = 'progress';
     }
+  }
+
+  ngOnChanges(changes: SimpleChanges) {
+    // Avoiding multiple calls for the same input in case there's no changes in the fields
+    // Checking if the fields that require validation are available and if they were changed
+    // In case were not changed, we stop the function. Otherwise, `validateInputValues` will be called.
+    if (
+      ['count', 'animation'].find(
+        key =>
+          changes[key] && (changes[key].isFirstChange() || changes[key].previousValue === changes[key].currentValue),
+      )
+    ) {
+      return;
+    }
+
+    this.validateInputValues();
   }
 
   ngAfterViewInit() {