+++ src/app/app.component.ts
@@ -11,6 +11,7 @@ export class AppComponent implements OnInit, OnDestroy {
   animation = 'pulse';
   contentLoaded = false;
   count = 2;
+  widthHeightSizeInPixels = 50;
 
   intervalId: number | null = null;
 
@@ -22,6 +23,7 @@ export class AppComponent implements OnInit, OnDestroy {
     this.intervalId = window.setInterval(() => {
       this.animation = this.animation === 'pulse' ? 'progress-dark' : 'pulse';
       this.count = this.count === 2 ? 5 : 2;
+      this.widthHeightSizeInPixels = this.widthHeightSizeInPixels === 50 ? 100 : 50;
     }, 5000);
   }